refactor(collections): split predicates into collection-filters; harden slug integrity
- Move the predicate registry (FILTERS + helper builders) into a new
  collection-filters.ts, leaving collections.ts as the join + validation + query
  API. collections.ts drops from ~400 to ~190 lines.
- Derive a CollectionSlug union from FILTERS (keyof typeof, via satisfies) and
  type the *_SLUGS grouping lists against it, so a slug typo in code is now a
  compile error.
- Validate the json <-> predicate join BOTH ways at module load: a json entry
  with no predicate, or a predicate with no json entry, throws — so the two
  files can never silently drift.
